requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery for background tasks
celery==5.3.6
flower==2.0.1

# MinHash & LSH
datasketch==1.6.4
mmh3==4.1.0

# Vietnamese NLP
underthesea==6.7.0

# PDF Processing
PyMuPDF==1.23.21
python-docx==1.1.0
pdf2image==1.17.0
pytesseract==0.3.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP & API
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# File Storage
boto3==1.34.34
minio==7.2.3

# Utilities
python-magic==0.4.27
pillow==10.2.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
ruff==0.1.14
mypy==1.8.0
black==24.1.1
